Moved model from homepage.go to correct models file
