+++ models/FormattedLog.go
@@ -1,0 +1,10 @@
+package models
+
+type FormatedLog struct {
+	RowClass                 string
+	TimeStampFormatted       string
+	HostName                 string
+	FileName                 string
+	TimeStampClosedFormatted string
+	TimeFileWasOpened        string
+}
